Throw DefinitionException on missing return type relation class, rename test util
